adding privacypolicyfile , adding feature of sending notifications to admin and subadmins when new user sign up , turninng delivery to sub_admin

+++ routes/api.php
@@ -27,22 +27,25 @@ Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanc
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/categories/products', [CategoryController::class, 'categoriesWithProducts']);
     Route::get('/categories', [CategoryController::class, 'index']);
-    Route::post('/categories', [CategoryController::class, 'store'])->middleware('role:admin');
+    // sub_admin has full CRUD on categories
+    Route::post('/categories', [CategoryController::class, 'store'])->middleware('role:admin,sub_admin');
     Route::get('/categories/{category}', [CategoryController::class, 'show']);
-    Route::put('/categories/{category}', [CategoryController::class, 'update'])->middleware('role:admin'); 
-    Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->middleware('role:admin');
+    Route::put('/categories/{category}', [CategoryController::class, 'update'])->middleware('role:admin,sub_admin');
+    Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->middleware('role:admin,sub_admin');
 });
 
 
 
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/products', [ProductsController::class, 'index']);
+    // Only admin can add new products or import
     Route::post('/products', [ProductsController::class, 'store'])->middleware('role:admin');
     Route::post('/products/import', [ProductsController::class, 'import'])->middleware('role:admin');
     Route::get('/products/{product}', [ProductsController::class, 'show']);
-    Route::put('/products/{product}', [ProductsController::class, 'update'])->middleware('role:admin'); 
-    Route::delete('/products/{product}', [ProductsController::class, 'destroy'])->middleware('role:admin');
-    Route::put('/products/{product}/change-status', [ProductsController::class, 'changeProductStatus'])->middleware('role:admin');
+    // sub_admin can edit and delete products
+    Route::put('/products/{product}', [ProductsController::class, 'update'])->middleware('role:admin,sub_admin');
+    Route::delete('/products/{product}', [ProductsController::class, 'destroy'])->middleware('role:admin,sub_admin');
+    Route::put('/products/{product}/change-status', [ProductsController::class, 'changeProductStatus'])->middleware('role:admin,sub_admin');
 });
 
 
@@ -51,7 +54,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/targets', [TargetController::class, 'index']);
     Route::post('/targets', [TargetController::class, 'store'])->middleware('role:admin');
     Route::get('/targets/{target}', [TargetController::class, 'show']);
-    Route::put('/targets/{target}', [TargetController::class, 'update'])->middleware('role:admin'); 
+    Route::put('/targets/{target}', [TargetController::class, 'update'])->middleware('role:admin');
     Route::delete('/targets/{target}', [TargetController::class, 'destroy'])->middleware('role:admin');
 });
 
@@ -59,7 +62,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/monthly-targets', [MonthlyTargetController::class, 'index']);
     Route::post('/monthly-targets', [MonthlyTargetController::class, 'store'])->middleware('role:admin');
     Route::get('/monthly-targets/{monthly_target}', [MonthlyTargetController::class, 'show']);
-    Route::put('/monthly-targets/{monthly_target}', [MonthlyTargetController::class, 'update'])->middleware('role:admin'); 
+    Route::put('/monthly-targets/{monthly_target}', [MonthlyTargetController::class, 'update'])->middleware('role:admin');
     Route::delete('/monthly-targets/{monthly_target}', [MonthlyTargetController::class, 'destroy'])->middleware('role:admin');
 });
 
@@ -67,7 +70,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/user-targets', [UserTargetController::class, 'index']);
     Route::post('/user-targets', [UserTargetController::class, 'store'])->middleware('role:admin');
     Route::get('/user-targets/{user_target}', [UserTargetController::class, 'show']);
-    Route::put('/user-targets/{user_target}', [UserTargetController::class, 'update'])->middleware('role:admin'); 
+    Route::put('/user-targets/{user_target}', [UserTargetController::class, 'update'])->middleware('role:admin');
     Route::delete('/user-targets/{user_target}', [UserTargetController::class, 'destroy'])->middleware('role:admin');
 });
 
@@ -75,7 +78,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/user-monthly-targets', [UserMonthlyTargetController::class, 'index']);
     Route::post('/user-monthly-targets', [UserMonthlyTargetController::class, 'store'])->middleware('role:admin');
     Route::get('/user-monthly-targets/{user_monthly_target}', [UserMonthlyTargetController::class, 'show']);
-    Route::put('/user-monthly-targets/{user_monthly_target}', [UserMonthlyTargetController::class, 'update'])->middleware('role:admin'); 
+    Route::put('/user-monthly-targets/{user_monthly_target}', [UserMonthlyTargetController::class, 'update'])->middleware('role:admin');
     Route::delete('/user-monthly-targets/{user_monthly_target}', [UserMonthlyTargetController::class, 'destroy'])->middleware('role:admin');
 });
 
@@ -91,7 +94,8 @@ Route::middleware('auth:sanctum')->group(function () {
 
 
 Route::middleware('auth:sanctum')->group(function () {
-    Route::get('/orders/all', [OrdersController::class, 'getAllOrders'])->middleware('role:admin,delivery');
+    // sub_admin can view all orders (but not necessarily change status - handled in controller)
+    Route::get('/orders/all', [OrdersController::class, 'getAllOrders'])->middleware('role:admin,sub_admin');
     Route::get('/orders/my-orders', [OrdersController::class, 'getMyOrders']);
     Route::post('/orders', [OrdersController::class, 'createOrder']);
     Route::get('/orders/{order}', [OrdersController::class, 'getSingleOrder']);
@@ -108,11 +112,13 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/profile', [UserController::class, 'myProfile']);
     Route::put('/profile/update', [UserController::class, 'updateProfile']);
     Route::put('/profile/fcm-token', [UserController::class, 'updateFcmToken']);
-    Route::get('/users/customers', [UserController::class, 'showCustomers'])->middleware('role:admin');
-    Route::get('/users/delivery-boys', [UserController::class, 'showDeliveryBoys'])->middleware('role:admin');
-    Route::post('/users/delivery-boy', [UserController::class, 'storeDeliveryBoy'])->middleware('role:admin');
-    Route::put('/users/delivery-boy/{delivery_boy}', [UserController::class, 'updateDeliveryBoy'])->middleware('role:admin');
-    Route::delete('/users/delivery-boy/{delivery_boy}', [UserController::class, 'deleteDeliveryBoy'])->middleware('role:admin');
+    // sub_admin can view customers
+    Route::get('/users/customers', [UserController::class, 'showCustomers'])->middleware('role:admin,sub_admin');
+    // Only admin manages sub-admins
+    Route::get('/users/sub-admins', [UserController::class, 'showSubAdmins'])->middleware('role:admin');
+    Route::post('/users/sub-admin', [UserController::class, 'storeSubAdmin'])->middleware('role:admin');
+    Route::put('/users/sub-admin/{sub_admin}', [UserController::class, 'updateSubAdmin'])->middleware('role:admin');
+    Route::delete('/users/sub-admin/{sub_admin}', [UserController::class, 'deleteSubAdmin'])->middleware('role:admin');
 });
 
 
@@ -129,7 +135,8 @@ Route::middleware('auth:sanctum')->group(function () {
 
 Route::middleware('auth:sanctum')->group(function () {
     Route::get('/offers', [OffersController::class, 'index']);
-    Route::post('/offers', [OffersController::class, 'store'])->middleware('role:admin');
-    Route::delete('/offers/{offer}', [OffersController::class, 'destroy'])->middleware('role:admin');
+    // sub_admin can manage offers
+    Route::post('/offers', [OffersController::class, 'store'])->middleware('role:admin,sub_admin');
+    Route::delete('/offers/{offer}', [OffersController::class, 'destroy'])->middleware('role:admin,sub_admin');
     Route::get('/offers/{offer}', [OffersController::class, 'show']);
 });
